Move WebAuthnPlugin definition to api::plugin
